+++ plugins/shape/rectangle/shrectangle.cpp
@@ -94,7 +94,7 @@ Plugin::Plugin() { }
 
 Plugin::~Plugin() { }
 
-int Plugin::type() const { return static_cast<int>(GiType::ShRectangle); }
+int Plugin::type() const { return GiType::ShRectangle; }
 
 QIcon Plugin::icon() const { return QIcon::fromTheme("draw-rectangle"); }
